Replaced ras2inundation.py code with newer geocurves-centric code and deleted produce_inundation_with_from_geocurves.py
